+++ Lottery/module.php
@@ -16,8 +16,8 @@ $lottery = \Modules\Lottery\Models\Lottery::query()
 ?&gt;
 {{ plural($lottery-&gt;amount, setting('moneyname')) }}</code></pre>
 INFO,
-    'version'  => '1.0.0',
-    'requires' => '14.0.0',
+    'version'  => '1.0.1',
+    'requires' => '14.0.3',
     'author'   => 'Vantuz',
     'email'    => 'admin@visavi.net',