+++ Notebook/Models/Notebook.php
@@ -6,6 +6,7 @@ namespace Modules\Notebook\Models;
 
 use App\Casts\HtmlCast;
 use App\Models\User;
+use Carbon\CarbonImmutable;
 use Illuminate\Database\Eloquent\Model;
 use Illuminate\Database\Eloquent\Relations\BelongsTo;
 use Illuminate\Support\HtmlString;
@@ -13,18 +14,14 @@ use Illuminate\Support\HtmlString;
 /**
  * Class Notebook
  *
- * @property int    $id
- * @property int    $user_id
- * @property string $text
- * @property int    $created_at
+ * @property int                  $id
+ * @property int                  $user_id
+ * @property string               $text
+ * @property CarbonImmutable      $created_at
+ * @property CarbonImmutable|null $updated_at
  */
 class Notebook extends Model
 {
-    /**
-     * Indicates if the model should be timestamped.
-     */
-    public $timestamps = false;
-
     /**
      * The attributes that aren't mass assignable.
      */