+++ Notebook/module.php
@@ -3,8 +3,8 @@
 return [
     'name'        => 'Блокнот',
     'description' => 'Личные заметки пользователя',
-    'version'     => '1.0.0',
-    'requires'    => '14.0.0',
+    'version'     => '1.0.1',
+    'requires'    => '14.0.3',
     'author'      => 'Vantuz',
     'email'       => 'admin@visavi.net',
     'homepage'    => 'https://visavi.net',