+++ Offer/module.php
@@ -6,8 +6,8 @@ use Modules\Offer\Models\Offer;
 return [
     'name'        => 'Предложения и проблемы',
     'description' => 'Модуль предложений и проблем',
-    'version'     => '1.0.2',
-    'requires'    => '14.0.1',
+    'version'     => '1.0.3',
+    'requires'    => '14.0.3',
     'author'      => 'Vantuz',
     'email'       => 'admin@visavi.net',
     'homepage'    => 'https://visavi.net',