+++ SocialAuth/module.php
@@ -3,8 +3,8 @@
 return [
     'name'        => 'Социальная авторизация',
     'description' => 'Авторизация через социальные сети (Google, GitHub, Yandex, VK)',
-    'version'     => '1.0.1',
-    'requires'    => '14.0.0',
+    'version'     => '1.0.2',
+    'requires'    => '14.0.3',
     'author'      => 'Vantuz',
     'email'       => 'admin@visavi.net',
     'homepage'    => 'https://visavi.net',